Simplify DatabaseConnection
Remove the caching of prepared statements and of the connection and just
use @ApplicationScoped, this simplifies the code quite a bit.

The caching of prepared statements was not useful at all (there are
currently no cases where the same query is used multiple times in the
same transaction). And getting the connection again for every query is
also fine, the JTA data source will return the "same" connection every
time during a transaction.