Stop using JsonObject and JsonArray and the builders
Instead use JsonMap and JsonList which are just normal mutable Maps and
Lists, so they can be used for building as well. Also use simple java
"primitive" types instead of the other JsonValue types.

This avoids unnecessary object allocations and copying (garbage) and is
sometimes more flexible to work with.